--- tests/000-testsMain.cpp
@@ -1,11 +1,0 @@
-//000- prefix is for the main-file to always be on top in folder view
-
-//use the main function provided by catch2
-#define CATCH_CONFIG_MAIN
-//include the catch2 library
-#include "./catch2/catch.hpp"
-
-//!!! following lines not nesseccary in the current test pipeline; all .cpp files are searched and compiled anyway
-
-//include the tests
-//#include "./initialize.h"
