+++ tests/catch2/catchMain.cpp
@@ -1,0 +1,6 @@
+
+//use the main function provided by catch2
+#define CATCH_CONFIG_MAIN
+
+//include the catch2 library
+#include "./catch.hpp"
